requirements.txt
piexif==1.1.*
Pillow==9.4.*